chore: add data to defer
