+++ apps/builder/src/page/AI/AIAgentRun/AIAgentRunPC/index.tsx
@@ -2,7 +2,7 @@ import { FC, useMemo, useState } from "react"
 import { Controller, useForm, useFormState } from "react-hook-form"
 import { useTranslation } from "react-i18next"
 import { useSelector } from "react-redux"
-import { useLoaderData } from "react-router-dom"
+import { useAsyncValue } from "react-router-dom"
 import { v4 } from "uuid"
 import {
   Button,
@@ -60,7 +60,7 @@ import {
 } from "./style"
 
 export const AIAgentRunPC: FC = () => {
-  const { agent, marketplaceInfo } = useLoaderData() as {
+  const { agent, marketplaceInfo } = useAsyncValue() as {
     agent: Agent
     marketplaceInfo: MarketAiAgent | undefined
   }
