+++ apps/builder/src/router/loader/agentRunLoader.ts
@@ -1,22 +1,24 @@
-import { LoaderFunction, redirect } from "react-router-dom"
+import { LoaderFunction, defer, redirect } from "react-router-dom"
 import { fetchAgentDetail, getAIAgentMarketplaceInfo } from "@/services/agent"
+
+const fetchAgentFullData = async (agentId: string) => {
+  const agent = await fetchAgentDetail(agentId)
+  let marketInfo = undefined
+  if (agent.data.publishedToMarketplace) {
+    marketInfo = await getAIAgentMarketplaceInfo(agentId)
+  }
+  return {
+    agent: agent.data,
+    marketplaceInfo: marketInfo ? marketInfo.data : undefined,
+  }
+}
 
 export const agentRunLoader: LoaderFunction = async (args) => {
   const { agentId } = args.params
   if (agentId) {
-    try {
-      const agent = await fetchAgentDetail(agentId)
-      let marketInfo = undefined
-      if (agent.data.publishedToMarketplace) {
-        marketInfo = await getAIAgentMarketplaceInfo(agentId)
-      }
-      return {
-        agent: agent.data,
-        marketplaceInfo: marketInfo ? marketInfo.data : undefined,
-      }
-    } catch (e) {
-      return redirect("/404")
-    }
+    return defer({
+      data: fetchAgentFullData(agentId),
+    })
   } else {
     redirect("/404")
   }
